+++ .github/workflows/job-server-test-coverage.yml
@@ -39,6 +39,8 @@ jobs:
           path: |
             server/internal/api/openapi_gen.go
             server/internal/book/*_gen.go
+            server/internal/bookcode/*_gen.go
+            server/internal/books/*_gen.go
             server/internal/user/*_gen.go
           key: generated-code-${{ github.sha }}
