+++ .github/workflows/job-server-test-medium.yml
@@ -38,6 +38,8 @@ jobs:
           path: |
             server/internal/api/openapi_gen.go
             server/internal/book/*_gen.go
+            server/internal/bookcode/*_gen.go
+            server/internal/books/*_gen.go
             server/internal/user/*_gen.go
           key: generated-code-${{ github.sha }}
