+++ server/internal/bookcode/create_book_by_code_service_test.go
@@ -1,13 +1,45 @@
 //go:build medium
 
-package book
+package bookcode
 
 import (
 	"context"
+	"database/sql"
 	"testing"
 
-	"holocron/internal/book/domain"
+	book "holocron/internal/book/domain"
+	"holocron/internal/bookcode/domain"
+
+	_ "github.com/mattn/go-sqlite3"
 )
+
+func setupTestDB(t *testing.T) *sql.DB {
+	t.Helper()
+	db, err := sql.Open("sqlite3", ":memory:")
+	if err != nil {
+		t.Fatalf("failed to open database: %v", err)
+	}
+	_, err = db.Exec(`
+		CREATE TABLE book_events (
+			event_id TEXT PRIMARY KEY,
+			book_id TEXT NOT NULL,
+			event_type TEXT NOT NULL,
+			code TEXT,
+			title TEXT,
+			authors TEXT,
+			publisher TEXT,
+			published_date TEXT,
+			thumbnail_url TEXT,
+			occurred_at TEXT NOT NULL
+		);
+		CREATE INDEX idx_book_events_book_id ON book_events(book_id);
+	`)
+	if err != nil {
+		t.Fatalf("failed to create table: %v", err)
+	}
+	t.Cleanup(func() { db.Close() })
+	return db
+}
 
 func TestCreateBookByCode_WithValidCode_ReturnsBook(t *testing.T) {
 	db := setupTestDB(t)
@@ -22,8 +54,8 @@ func TestCreateBookByCode_WithValidCode_ReturnsBook(t *testing.T) {
 	}
 	sources := []domain.BookInfoSource{
 		DBCacheSource(queries),
-		ExternalAPISource(googleFetcher.Fetch, domain.BookInfoFromGoogleBooks),
-		ExternalAPISource(openBDFetcher.Fetch, domain.BookInfoFromOpenBD),
+		ExternalAPISource(googleFetcher.Fetch, book.BookInfoFromGoogleBooks),
+		ExternalAPISource(openBDFetcher.Fetch, book.BookInfoFromOpenBD),
 	}
 
 	output, err := CreateBookByCode(context.Background(), queries, sources, CreateBookByCodeInput{
@@ -64,8 +96,8 @@ func TestCreateBookByCode_WithSameCode_ReturnsBookWithSameInfo(t *testing.T) {
 	}
 	sources := []domain.BookInfoSource{
 		DBCacheSource(queries),
-		ExternalAPISource(googleFetcher.Fetch, domain.BookInfoFromGoogleBooks),
-		ExternalAPISource(openBDFetcher.Fetch, domain.BookInfoFromOpenBD),
+		ExternalAPISource(googleFetcher.Fetch, book.BookInfoFromGoogleBooks),
+		ExternalAPISource(openBDFetcher.Fetch, book.BookInfoFromOpenBD),
 	}
 	first, _ := CreateBookByCode(context.Background(), queries, sources, CreateBookByCodeInput{
 		Code: "9784873115658",
@@ -85,4 +117,3 @@ func TestCreateBookByCode_WithSameCode_ReturnsBookWithSameInfo(t *testing.T) {
 		t.Errorf("expected same title %q, got %q", first.Title, second.Title)
 	}
 }
-
