+++ server/internal/books/create_book_service_test.go
@@ -1,6 +1,6 @@
 //go:build medium
 
-package book
+package books
 
 import (
 	"context"
