+++ server/internal/books/domain/book_item.go
