+++ server/internal/books/domain/book_item_test.go
