+++ server/internal/books/list_all_books_source.go
@@ -1,9 +1,9 @@
-package book
+package books
 
 import (
 	"context"
 
-	"holocron/internal/book/domain"
+	"holocron/internal/books/domain"
 )
 
 func ListAllBooksSource(queries *Queries) domain.BookListSource {
