+++ server/internal/books/list_books_service.go
@@ -1,9 +1,9 @@
-package book
+package books
 
 import (
 	"context"
 
-	"holocron/internal/book/domain"
+	"holocron/internal/books/domain"
 )
 
 type ListBooksInput struct {
